requirements/base.txt
Django==3.2.3
django-admin-interface==0.16.3
djangorestframework==3.12.4
djangorestframework-gis==0.17
djangorestframework-simplejwt==4.4.0
django-filter==2.3.0
docutils==0.14
drf-yasg==1.20
pyyaml==5.2
pyjwt==1.7.1
psycopg2-binary==2.8.2
requests==2.19.1